# Conflicts:
#	frontend/src/views/resource/resource/index.vue

+++ backend/backend-admin/src/main/java/com/qwq/user/service/IStudentService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface IStudentService {
     List<Student> listStudent(Student student);
-    int LearningTime(Long studentId);
+    Long LearningTime(Long studentId);
 }
